Move sanitize_url for mypy
